+++ exonum/src/storage/sparse_list_index.rs
@@ -21,7 +21,11 @@ use byteorder::{BigEndian, ByteOrder};
 
 use std::{borrow::Cow, cell::Cell, marker::PhantomData};
 
-use super::{indexes_metadata::IndexType, BaseIndex, BaseIndexIter, Fork, Snapshot, StorageKey,
+use super::{base_index::{BaseIndex, BaseIndexIter},
+            indexes_metadata::IndexType,
+            Fork,
+            Snapshot,
+            StorageKey,
             StorageValue};
 use crypto::{hash, CryptoHash, Hash};